requirements.txt
mpi4py==4.1.1
opencv-python==4.10.0.84
pytesseract==0.3.13
python-Levenshtein==0.27.3
numpy>=1.26.2
pandas>=2.0.0